Struct cge::network::Network

pub struct Network {
    pub size: usize,
    pub genome: Vec<Gene>,
}

Fields

size

genome

Methods

impl Network

fn evaluate(&mut self, inputs: &Vec<f64>) -> Vec<f64>

Evaluates the neural network with the given inputs, returning a vector of outputs. The encoding can encode recurrent connections and bias inputs, so an internal state is used. It is important to run the clear_state method before calling evaluate again, unless it is desired to allow data carry over from the previous evaluation, for example if the network is being used as a real time controller.

If too little inputs are given, the empty inputs will have a value of zero. If too many inputs are given, the extras are discarded.

Examples

use cge::Network;

// Load a neural network
let mut network = Network::load_from_file("neural_network.ann").unwrap();

// Get the output of the neural network the the specified inputs
let result = network.evaluate(&vec![1.0, 1.0]);

// Get the output of the neural network with no inputs
let result = network.evaluate(&Vec::new());

// Get the output of the neural network with too many inputs (extras aren't used)
let result = network.evaluate(&vec![1.0, 1.0, 1.0]);
 
// Let's say adder.ann is a file with a neural network with recurrent connections, used for
// adding numbers together.
let mut adder = Network::load_from_file("adder.ann").unwrap();

// result_one will be 1.0
let result_one = adder.evaluate(&vec![1.0]);

// result_two will be 3.0
let result_two = adder.evaluate(&vec![2.0]);

// result_three will be 5.0
let result_three = adder.evaluate(&vec![2.0]);

// If this behavior is not desired, call the clear_state method between evaluations:
let result_one = adder.evaluate(&vec![1.0]);

adder.clear_state();

// The 1.0 from the previous call is gone, so result_two will be 2.0
let result_two = adder.evaluate(&vec![2.0]);

fn clear_state(&mut self)

Clears the internal state of the neural network.

fn save_to_file(&self, path: &str) -> Result<()>

Saves the neural network to a file. Returns an empty tuple on success, or an io error.

Examples

use cge::Network;

// Create a neural network
let network = Network {
    size: 0,
    genome: Vec::new()
};

// Save network to neural_network.ann
network.save_to_file("neural_network.ann");

fn load_from_file(path: &str) -> Result<Network>

Loads a neural network from a file. No guarantees are made about the validity of the genome. Returns the network, or an io error. If the file is in a bad format, io::ErrorKind::InvalidData is returned.

Examples

use cge::Network;

// Loads a network from the file neural_network.ann
let mut network = Network::load_from_file("neural_network.ann").unwrap();

fn get_subnetwork_index(&self, id: usize) -> Option<Range<usize>>

Returns the start and end index of the subnetwork starting at the neuron with the given id, or None if it does not exist.

fn get_neuron_index(&self, id: usize) -> Option<usize>

Returns the index of the neuron with the given id, or None if it does not exist.